Moved the FEN String tests into FENStringTests.java and moved the moves testing if moves are legal into LegalMovesTests.java
